% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{CASCR}
\alias{CASCR}
\title{Semiparametric Causal Mediation Modeling of Semi-Competing Risks}
\usage{
CASCR(
  df,
  effect = c("DE", "IE"),
  intervention = c(1, 0),
  cal_level = "median",
  myunit = "raw",
  downsample = 1,
  sen_ana = FALSE,
  get_variance = c("asymptotic"),
  boot_times = 1000,
  timer = TRUE,
  num_of_cores = 1,
  plot_result = FALSE,
  variance_method = "new",
  threshold = 1e-12
)
}
\arguments{
\item{df}{a data frame with designated column order: \code{T1}, \code{T2}, \code{d1}, \code{d2}, \code{Z}, \code{X} where X maybe empty or a matrix.}

\item{effect}{the causal effect to be estimated. Choices are \code{"DE"} or \code{"IE"}. Default is \code{c("DE", "IE")}}

\item{intervention}{numeric vector with two elements. Default is \code{c(1, 0)}.}

\item{cal_level}{the level \code{X} should be evaluated at. It can be a numeric vector with \code{length(cal_level) = dim(X)[2]} or \code{"median"}, \code{"mean"}, etc. Default is \code{"median"}.}

\item{myunit}{the length to be considered as one unit. Default is \code{"raw"}.}

\item{downsample}{indicates how many consecutive event should be considered as one event. Default is \code{1}.}

\item{sen_ana}{doing sensitivity analysis or not. Default is \code{FALSE}.}

\item{get_variance}{the method to compute the variance. Choices are \code{"a"} or \code{"b"}. Default is \code{"a"}.}

\item{boot_times}{the times of bootstrap. Default is \code{1000}.}

\item{timer}{will show the progress. Default is \code{TRUE}.}

\item{num_of_cores}{the number of cores assigned. Default is \code{1}.}

\item{plot_result}{will show some primary result. Default is \code{FALSE}.}

\item{threshold}{specifies if a logistic regression converge or not. Default is \code{1e-12}.}
}
\value{
\code{CASCR} returns a list with components specified by \code{effect}.
}
\description{
This function analysis semicompeting risks data and gives the estimators of direct and indirect effects, along with their variances.
}
